requirements.txt
Django==6.0.1
djangorestframework==3.16.1
psycopg2==2.9.11
python-dotenv==1.2.1
celery==5.6.2
redis==7.1.0
django-celery-beat==2.8.1
requests==2.32.5
django-cors-headers==4.9.0
drf-yasg==1.21.14